Remove API since it won't be ported to 1.16 and move to local capture
